set led feedback sensitivity

+++ A2Z-CAR-F7/A2Z-F7-Project/SW4STM32/STM32746G_DISCOVERY/Application/User/lcd_user_interface.h
@@ -15,6 +15,6 @@
 #include "lcd_log.h"
 
 void draw_background();
-void draw_sensor_data(int sensor_num, uint16_t radius);
+void draw_sensor_data(int sensor_num, uint8_t radius);
 
 #endif /* APPLICATION_USER_LCD_USER_INTERFACE_H_ */
